Make xminus module available as window.xm on mount
The generated code needs to access some functions from the module at
runtime. expoing them on window looks hacky. Insead we can just import
ourselves (which just gives us a reference to the module and nothing else) and
expose that reference on the window.

We want a short id. mithril uses m, so let's be copycats and use xm.