Add support for importing components
as we want to support imports inside the component files we need to use dynamic
import rather than eval - and change turn any relative import urls inside the
component module into absolute urls. That means no support for dynamic import
from component files (which isn't a loss for me) in return for not requiring a
bundle step.

There's more problems with this approach (e.g. the components defined in the
module that calls mount are not linked to their hook objects because neither
import nor eval give us access to that scope). But whatever - right now I want
to explore my options.