Return array of nodes from both component and closure
For <template> nodes we need to grab from content.childNodes. Closures around a
template can also be expected to only care about node.childNodes while other
closures might want the complete node.

All of this is in preparation for unifying the whole append/replace/remove node
shebang. We wouldn't have to if it wasn't for DocumentFragments. As document
fragments 1. do not have persistent children (i.e. cannot be appended multiple
times) and 2. are not positioned (they don't have a position in the document
when they have 0 children) we have to do the housekeeping ourselves and wrap
those methods. It's nice [1] to know others are sad about this as well
[2].

[1] https://xkcd.com/979/
[2] https://github.com/whatwg/dom/issues/736