Introduce minimal cli: live reload server + headless test wrapper
Real basic cli for webdev - serve with livereload and run headless. Eventually
I'll also add a build target. Maybe more.

I wanted livereload and an automated test runner. Both require watching a given
path for changes. As inotify is a pain to work with [1], fanotify is not there
yet [2] and the whole idea of xminus is small projects (i.e. not a lot of files
to walk through) I went with polling by traversing the path on a timer. To make
things even simpler to implement I'm not even checking modification times of
files against themselves but rather add all of them up and treat any difference
to the previous sum as a change. Not perfect - but good
enough. Overflow/wraparound doesn't matter (different sum is different) and
edge cases where changes don't change the sum (to lazy to think about it - i
guess it's possible) doesn't matter either because we don't need perfection for
livereloading. For now that timer is set to 500ms because it's a nice round
number. we're also limited by the page reload time of the browser - 500ms
seems like a good compromise between cpu load, browser refresh time and
realtime-feeling for me. Whatever.

[1] Recursive file watching in linux in general - i hear other OS's fare better
in that regard. Maybe next year. I hear it will be the year of the linux
desktop.

[2] fanotify supports recursive watches for a path / bind - but only for
non-directory (e.g. move) events - I've read those should be supported in newer
kernel versions but I couldn't get them working (only create|write).
Apart from that it requires cap_sys_admin (i.e. either setcap or sudo).
See fanotify_init and fanotify_mark - especially the docs on FAN_MARK_MOUNT
https://man7.org/linux/man-pages/man2/fanotify_mark.2.html