Generate reference to node for access to node rather than DOM position
vnode.node is only guaranteed to be the node corresponding to vnode in create -
once we reach update it is possible that the vnode.node variable contains a
different value - e.g. a placeholder. If we want to access the node rather than
the position in the DOM (e.g. firstChild, nextSibling, ...) we need to create
our own reference.