Add bind macro
Setting the binding every update swallows input when the value changed in
between the scheduling of the update and the update. Just setting the binding
when the property is empty means we don't try to do the browsers job of keeping
the input value updated. This can't be everyhing but it solves my current
problem so yagni.