Refactor compile: Prettier html for static (!dataUrl) compilation
Compiling to data urls is necessary for runtime compilation as we want to await
the module loading and can't do that with script elements - dataUrls alllow us
to use `await import(dataUrl)`.

Static compilation on the other hand can use simple script elements - data urls
just unnecessarily obfuscate the generated code.

While I was at it I reevaluated my choice of using evaling component setup code
in the scope of the x-module - and linking the component to it's hooks object
by name. I don't like it anymore - was a fun idea but is unnecessarily complex
- it's much more obvious to link hooks via a xm.hooks object. That allows for a
lot of other places to be refactored and most importantly get's rid of the
requirement of splitting mounting and hook code (module and x-module) in the
index file. It also paves the way for multiple modules per file (no need to
restrict to one anymore). Apart from that there's a lot of things I noticed on
the way and because this is a solo project I don't care about clean diffs that
much.

Figuring out how to provide access to xm.components without passing the url for
the module to the component setup code was fun. As we only know it in mount (at
runtime) and need it sync (so it's available in all scripts following the
mounting one) we expose only the components first (we have access to those) and
then later replace window.xm with the dynamic import of ourselves. It's nice
that dynamic self imports are possible. While it would be nicer if there was a
self reference from the beginning, this is still much nicer than keeping a list
of exports to expose on the window.

We also switch to producing html rather than xml - I only went with xml because
outerHTML does not include the doctype. xml had edgecases (e.g. html inside
script tags is html encoded) so i shouldn't have made that tradeoff in the
first place.

Also only data urls need to (and can be without problem) complete urls - for
everything else we just need complete paths (and complete urls are actually bad
because they include the script runners host and port rather than leaving that
open to the serving environment).

I think there's more and I didn't proof read because it's late. yolo.