Unify component calling and add more $internal vars (~routing)
Until now we've been passing $update as a special internal var. For routing we
want 2 more - path and query params. So let's replace $update with
those. There's a little bit of bookkeeping necessary for syncing the vars with
the state of $internal but I want to explore this option - this is the start of
a tracer bullet for turning $ properties into dynamic variables. Or something
like that.