Improve hash router init: Set using replaceState, only on init
Use replaceState to not fuck with back button behavior coming from non hash
page (with location hash we effectively prevent back button to and beyond the
non-hash page).

We're not using the prettier /#/ because that fucks with relative links.

the relative url ./bar will mean different things in /foo/#/ (/foo/bar) and
/foo#/ (/bar).