Support modifiers in .on macro: Disable auto $update via "no"(update)
While automatic updating is good default, sometimes it's not what we want. Two
common cases are (1) async handlers (where the auto $update happens to early)
and (2) key handlers that only act on specific keys (e.g. enter) and don't do
anything and thus don't require an update on most calls.

Of the ways existing frameworks solve this [1] I prefer the modifier
approach svelte and vue take. It's very simple to implement and extensible -
multiple modifiers can be chained. While svelte opts for `|` I don't see the
need - this is more elegant to implement just as understandable imho.

I decided on "no" rather than "noupdate" as it will be used quite often and
I didn't like the verbosity of "noupdate" when I added it to the
templates. "no" should be clear enough and circumvents that.

[1]
- modifiers in svelte and vue
- state modification via setter function (i.e. update on modification) of react
  (not sure I'm representing this correctly, but whatever)
- special event property of mithril (event.redraw = false) to disable update