Examples: Refactor hackernews stories into table
It's a table. And we don't want it to be responsive and change the number of
columns. It should always be a table with 3 columns. I drank too much of the
web dev coolaid and didn't even consider using a table because tables bad / not
webscale enough. Well... it works. and it's much less fancy. I like it.

This is exactly why I started writing my own framework - to open my eyes to the
edge cases of the rules instilled by the community. Or maybe I'm talking
bullshit and this will break horribly - but now that I think about it I don't
remember reading reasons against tables that would apply in this case - just
blanket statements to never use them.

Also i want the point and comment columns to be smaller on mobile - shortening
he widest cell (header) is the naive way to get there and doesn't pts & cmnts
just scream hackernews? It's understandable and works. :shipit: