Examples: Visualize rank and via color
As we want visited links to be grey we can't extend the rank to comments via
font-color and have to use background color. Points on the other hand is not a
link so we can use it's font-color to visualize freshness - that goes
intuitively with font-color for visited.

Also refactored localStorage a little. Don't like that pattern much either yet
- but not sure where it's going yet either. So still waiting until I extract it
into something into xminus.