Make .if macro render and not only attach conditionally
Until now the if macro has been rendering the vnode in any case and only
conditionally toggling whether it gets attached to the DOM. While that's easier
to implement, performs better given frequent toggling and behaves the same in
most cases, it breaks when the vnode cannot be rendered when the condition is
not true (e.g. `<p .if=$.item>{$.item.title}</p>` throws on render when
`$.item` does not exist).

Changing the if macro to toggle rendering rather than only attaching is more
correct (and performs better when the condition is never met) even if more
verbose.

This change also required reusing existing internal vars in component update
rather than creating new ones in the update scope (should have done so from the
beginning, no drawbacks) and switching the `.on:update` to `.on:create` - which
is a breaking change but feels intuitive enough given our requirements for the
if macro (conditional rendering, i.e. re-creation).