Improve rebaseURL: Only rebase given basePath (absolutify otherwise)
When bundle is called dynamically we want to create a data url to import rather
than html with script tags as we want to await the import and that's easier
with a data url. This is configured by not setting a basePath. As
relative (even root relative) imports don't work inside script tags we need to
honor this in rebaseURL and only return rebased (root) relative urls when a
basePath is provided.