Global Refactor: Remove fragments and introduce custem elements
My hacky fragment wrapper was fundamentally broken and would have had to go
later on anyways [1]. Rather than just removing fragments and turning previous
fragments into divs I opted to pull in custom elements - they're just fancy
divs. This led to that and here we are - a lot changed.

- hook objects have been replaced with classes. I need to extend HTMLElement
  for custom elements at some point and there's no value in hiding that from the
  user
- Moving global variables $update, $path & $query into the global $ felt right
  now that it really can be used as the global state as local state can now be
  represented with the `this` of the enclosing component. Also properties was
  exhausting to write - props is just as expressive and shorter. With props
  comes slot for the children. Taken from what custom elements with shadow dom
  call those
- By pulling out most of the component setup code into the runtime we have less
  hard to debug and maintain generated code.
- Removed dynamic node tags for now because I'm not even sure I want them (I
  just wanted to keep that option - but I haven't used it yet) and they add
  quite a bit of complexity. Also they're not tested and I'm too lazy to write
  tests right now.

Notably, this commit does not introduce shadow dom. I experimented with that
but ended up deciding against it due to the huge amount of complexity they
entail.
1. query selector breaks because every component instance is their own
   document. integration tests become annoying because of that - there's no
   tooling in the stdlib and you have to build something out of treewalker. Not
   that much code (shitty and wrong but good enough for tests is <20LOC) but
   made me feel wrong.
2. Complete isolation means components do not inherit styles from their
   parent - but that's exactly what I want. I want some shared styles (e.g. css
   reset) and styles scoped to the component. I explicitly want scope
   inheritance; The thing that sucks about css to me is not inheritance, it's
   the lack of (nested) scopes. Shadow DOM means we have to clone every
   stylesheet we want in the component into it. That just didn't feel right to
   me.

I probably missed a lot, but I'm not in the mood to write more. I wish commit
messages were editable without rewriting all history. Something as close to the
code as commit messages but more like a wiki... Anyways.

[1] My hacky fragment wrapper only worked for create - it didn't handle any
modifications (i.e. update). If nodes changed, (re)moving the fragment only
moved the nodes that already existed at create - because any append would
happen on the current parent (i.e. a descendant of the document) and not the
document fragment.
I decided on removing the complexity of fragments rather than fixing the bugs
because (1) what I thought needed fragments for is actually achievable with
`display: contents` in 2020 (afaict) and (2) I wanted to add custom element
goodness (remove hook, scoped css, ???) in any case and just put it off because
I couldn't think of how to integrate it with fragments.