Revert Me: Fix component $ scope
I wanted to get rid of creating nested $ for components and removed it while
refactoring because we have access to `this` - but I don't know how exactly I
want to go about it (e.g. what about nested scope for for loops?). Let's
introduce it back for now to not break existing assumptions of nested $.