compiler: Refactor generateClosure and symbol-namespace runtime updates
Prevent method conflicts by using symbols [1] to remove edge cases (node that's
both for and if). This also allows us to dry up usages of generateClosure.

While we're at it we can also symbol-namespace the component update method to
be consistent.

[1] or rather: more specific names and symbols is just a freebie
at that point which additionally prevents conflicts with external code