test: Split throws into throws and rejects
splitting rejects out of throws allows us to ensure throws is called
sync (i.e. does not return a value / promise) unless the method name reminds
the user to await the call. pendingAssertions was a hack and i guess something
like this can work as long as tests are called in order and never
concurrent. maybe i'll revisit it - but for now i dislike the resulting loss of
early exit on error (unawaited promises will cause the following lines to
execute before the rejection). this way seems nicer