general: Only splat configured (x-props) props onto component instance
With splat props prop keys are dynamic and we'd have to clean up the previous
props assignment to not potentially end up with stale props assigned to the
instance. Rather than eating the runtime cost of that, let's move to assigning
only a configured list of props. This fixes both the stale state problem and
makes the external interface more explicit.

Having the props on the template is a first draft - not sure whether i prefer
that or insight the component class or even both. just something to get started