compiler: Make instantiation move top to bottom
With x-template props the approach of preparing the component while it's
disconnected and using connectedCallback to instantiate it once appended feels
even more hacky. It's much cleaner to make instantiation explicit and move it
where we want it - inside the generated create of the parent, where we have all
the data we need to instantiate.

As unconnected custom elements are not upgraded (i.e. are instances of
HTMLElement rather than Component) and elements inside templates (i.e. created
via template.innerHTML) do not count as connected, we have use
document.importNode to import and thus upgrade these nodes. Nodes created using
document.createElement count as connected.

As we're now executing the generated code independent of being appended to the
DOM we have to explicitly pass the parent for inject to keep working.

Also connectedCallback apparently can get called before the children of the
element (xSlot) have been parsed. Seems to be a whole thing. Hadn't noticed it
yet as a bug. init() should be less susceptible to the problem.