runtime: Allow munting multiple components by passing html
Sometimes we want to mount multiple components. Rather than exposing a create
function (easy enough to replicate - document.create + init) we allow passing
html to mount and thus allow using x-templating to create the initial
setup. As we're not interested in the x-mount element in terms of layout we set
it to display contents. Could do the same in base.css - not sure where the best
place is