runtime: Load x-dev content automatically via html body attribute
Keeping dev setup and tests in the same file as the production code requires
marking it for exclusion when bundling for production - we use the [x-dev]
attribute for that. As init applies to the document context into which xminus
is imported and bundling for production runs in an empty context / without
[x-dev] attributes anywhere we can setup the development environment for e.g. x
module files automatically when opening the file rather than loading it via
bundle.