test: Add t.log for pretty logging in tests
console.logs during tests are not indented and print as the test executes,
i.e. before the test name and the pass/fail line.

As we now have another variable we want to keep track of as current and
currentFn doesn't sound nice enough we're joining the state into the `current`
variable. I was close to going with currentFn but this feels cleaner in
retrospect and is the kind of refactoring i would value in projects at
work.