cli: Move test fixture writing into test module
we need js to serialize the json because assertJSON compares strings and
depends on equal property order - and that requires some hacky logic for
pretty indentation in go. now that we have cusom headless handlers we can just
move the logic to js.

not sure i like the POST pattern that much but can't be bothered to change it
right now. works well enough.