Use static esm imports for self reference and clean up runtime init
dynamic imports to get a reference to the current module are async - both
firefox and chrome are happy to do static self imports so let's try that for
simplicity. Then again, mount must be called async in any case (as the mount
call will be lexically before the custom element definitions when bundled and
app.init will thus not exist / not upgraded when called sync) so we have to
ensure mount is async using setTimeout when bundled and we don't bundle
dynamically.

also cleaned up some legacy leftovers