test: Collect before/after eachs at test runtime
we cannot collect the each wrappers inside group (`t.describe{,Only}`) because
not all tests are grouped - bare calls to `t.beforeEach` and `t.afterEach`
should be allowed and right now don't have any wrappers. this is the next
best place i could think of.

we need to slice (copy) array because `reverse` mutates the original.