app: Support back/forward scroll restoration by caching views
Browsers restore the scroll position on back forward navigation - sadly they
don't expose a hook to delay the restoration and component rendering is
async. If the scroll position to be restored is outside the default body height
(without the route component rendered) the wrong position is restored. By
caching the component for each route we ensure the component with the correct
previous height is in the body when the browser tries to restore the scroll
position. The scroll(0,0) is ignored on back/forward navigation (or happens
before the restoration and is thus overridden) - which is a nice coincidence as
there is (afaik) no clean way to check whether the current hashchange is a
back/forward navigation. The performance navigation api and pageshow event
don't seem to care about hash changes.

If we'd use server side routing we wouldn't have this problem because all major
browsers have a bfcache [1] that does the caching for us. That would require
creating an html page for each route and more and feels even more hacky than
this solution.

[1] https://web.dev/bfcache/