Component: Replace onCreate with onRender hook
Oftentimes I want to execute something on first render and every update - let's
try abstracting that instead of onCreate (which should rather be called
onCreated as it runs after and i haven't found a use case for yet). not sure
about the order but i want it to happen before any values are used (i.e. before
the compiled create/update).