freiluftkino: Make pwa manifest urls relative
turns out start_url is optional [1] - by ignoring it and using relative urls in
icons [2] we can hopefully install the pwa when served in dev mode (i.e. on `/`
rather than `/xminus/`). still haven't figured out how to do that with firefox
on android - assuming it's because the pwa is not served via https [3]. maybe
there's a trick

[1] https://developer.mozilla.org/en-US/docs/Web/Manifest/start_url
[2] https://developer.mozilla.org/en-US/docs/Web/Manifest/icons
[3] https://developer.mozilla.org/en-US/docs/Web/Progressive_web_apps/Installable_PWAs