cli: Add ssh port forwarding
PWA's can only be installed when served via https [1] - i.e. serving on 0.0.0.0
is not enough to install dev versions of the pwa on android. By forwarding the
local http server to a server we can serve everything over https.

We could do the same using the go crypto/ssh package but that's more lines of
code and should do just as nicely.

[1] https://developer.mozilla.org/en-US/docs/Web/Progressive_web_apps/Installable_PWAs