compiler: Fix inject for unattached components
Components are potentially init'ed before being attached to the document,
i.e. closest won't be able to find all their ancestors because they're not part
of the complete tree yet (e.g. inside nested for loops / generateClosure). As
we're only interested in the component parents and we're already storing those
we can just walk the tree ourselves.