icon: Vertically center icons
Apparently line-height does not apply to the svg icons - to make them look in
line with sibling text we need to center. By removing the height the x-icon
fills the line and the svg can be vertically centered inside of it (due to the
width of 1em and svgs respecting aspect ration the svg itself is 1em high).
No idea what i'm doing here, this was trial and error.