jsxy: Fix global hooks var mixups between parents and children
Given parent `x` with children `a` and `b`, rendering the children of `a` will
override the global hooks var (in each call to renderChildren) and cause b to
look into the hooks of `a` (or a child of `a`) instead of the hooks of `x`. We
have to reset the hooks var once we've finished calling renderChild.