frlkino: Fix bug when switching routes / main components
as refs.main would be swapped out when switching routes but not updated inside
the copy of refs we kept around in useEffect we'd try to update a node that
isn't attached anymore after the first route change. subscribing to changes
inside each component works. I guess a single subscription and setting button
state in html`` rather than useEffect is the way to go. whatever for now.