jsxy: Force <a> to render unchanged routes (match normal link behavior)
Clicking a link to the current hash (path/query) does not fire a
popstate nor hashchange event; non-hash links behave differently (they always
do something, even if that thing is rerendering the current page) and our hash
router should match that behavior.

Double hashing routes in the url ensures hash links never match the current url
(if we stick with single hashed hrefs) at the cost of a little url noise (and
sometimes a little url jumping from the hash modification). It's hacks on hacks
on hacks all the way down - maybe just one more...

The use case that brought this up: The hamburger menu overlay closes
automatically when rendering (as the default state is closed) - I didn't want
to add an edge case for when the user clicks on the current url to hide the
overlay. So here we are - everything's an edge case. Happy now?

Also fixing a bug with path params that I'm too lazy to add tests for right
now (sue me). /{a}/{b} didn't work before, now it should :TM:.